+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/CharTest.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class CharTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerCHAR
+     *
+     * @param mixed $expectedResult
+     * @param $character
+     */
+    public function testCHAR($expectedResult, $character)
+    {
+        $result = TextData::CHARACTER($character);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerCHAR()
+    {
+        return require 'data/Calculation/TextData/CHAR.php';
+    }
+}
