+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/CleanTest.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class CleanTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerCLEAN
+     *
+     * @param mixed $expectedResult
+     * @param $value
+     */
+    public function testCLEAN($expectedResult, $value)
+    {
+        $result = TextData::TRIMNONPRINTABLE($value);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerCLEAN()
+    {
+        return require 'data/Calculation/TextData/CLEAN.php';
+    }
+}
