+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/ConcatenateTest.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class ConcatenateTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerCONCATENATE
+     *
+     * @param mixed $expectedResult
+     */
+    public function testCONCATENATE($expectedResult, ...$args)
+    {
+        $result = TextData::CONCATENATE(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerCONCATENATE()
+    {
+        return require 'data/Calculation/TextData/CONCATENATE.php';
+    }
+}
