+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/DollarTest.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class DollarTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerDOLLAR
+     *
+     * @param mixed $expectedResult
+     */
+    public function testDOLLAR($expectedResult, ...$args)
+    {
+        $result = TextData::DOLLAR(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerDOLLAR()
+    {
+        return require 'data/Calculation/TextData/DOLLAR.php';
+    }
+}
