+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/ExactTest.php
@@ -1,0 +1,51 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class ExactTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerEXACT
+     *
+     * @param mixed $expectedResult
+     * @param array $args
+     */
+    public function testEXACT($expectedResult, ...$args)
+    {
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(' ');
+        StringHelper::setCurrencyCode('$');
+
+        $result = TextData::EXACT(...$args);
+        $this->assertSame($expectedResult, $result);
+    }
+
+    /**
+     * @return array
+     */
+    public function providerEXACT()
+    {
+        return require 'data/Calculation/TextData/EXACT.php';
+    }
+}
