+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/FindTest.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class FindTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerFIND
+     *
+     * @param mixed $expectedResult
+     */
+    public function testFIND($expectedResult, ...$args)
+    {
+        $result = TextData::SEARCHSENSITIVE(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerFIND()
+    {
+        return require 'data/Calculation/TextData/FIND.php';
+    }
+}
