+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/LeftTest.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class LeftTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerLEFT
+     *
+     * @param mixed $expectedResult
+     */
+    public function testLEFT($expectedResult, ...$args)
+    {
+        $result = TextData::LEFT(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerLEFT()
+    {
+        return require 'data/Calculation/TextData/LEFT.php';
+    }
+}
