+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/LowerTest.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class LowerTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerLOWER
+     *
+     * @param mixed $expectedResult
+     * @param $value
+     */
+    public function testLOWER($expectedResult, $value)
+    {
+        $result = TextData::LOWERCASE($value);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerLOWER()
+    {
+        return require 'data/Calculation/TextData/LOWER.php';
+    }
+}
