+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/ProperTest.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class ProperTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerPROPER
+     *
+     * @param mixed $expectedResult
+     * @param $value
+     */
+    public function testPROPER($expectedResult, $value)
+    {
+        $result = TextData::PROPERCASE($value);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerPROPER()
+    {
+        return require 'data/Calculation/TextData/PROPER.php';
+    }
+}
