+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/ReplaceTest.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class ReplaceTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerREPLACE
+     *
+     * @param mixed $expectedResult
+     */
+    public function testREPLACE($expectedResult, ...$args)
+    {
+        $result = TextData::REPLACE(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerREPLACE()
+    {
+        return require 'data/Calculation/TextData/REPLACE.php';
+    }
+}
