+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/SubstituteTest.php
@@ -1,0 +1,43 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class SubstituteTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerSUBSTITUTE
+     *
+     * @param mixed $expectedResult
+     */
+    public function testSUBSTITUTE($expectedResult, ...$args)
+    {
+        $result = TextData::SUBSTITUTE(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerSUBSTITUTE()
+    {
+        return require 'data/Calculation/TextData/SUBSTITUTE.php';
+    }
+}
