+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/TTest.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class TTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerT
+     *
+     * @param mixed $expectedResult
+     * @param $value
+     */
+    public function testT($expectedResult, $value)
+    {
+        $result = TextData::RETURNSTRING($value);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerT()
+    {
+        return require 'data/Calculation/TextData/T.php';
+    }
+}
