+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/TextJoinTest.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class TextJoinTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerTEXTJOIN
+     *
+     * @param mixed $expectedResult
+     * @param array $args
+     */
+    public function testTEXTJOIN($expectedResult, array $args)
+    {
+        $result = TextData::TEXTJOIN(...$args);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerTEXTJOIN()
+    {
+        return require 'data/Calculation/TextData/TEXTJOIN.php';
+    }
+}
