+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/TrimTest.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class TrimTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerTRIM
+     *
+     * @param mixed $expectedResult
+     * @param $character
+     */
+    public function testTRIM($expectedResult, $character)
+    {
+        $result = TextData::TRIMSPACES($character);
+        $this->assertEquals($expectedResult, $result);
+    }
+
+    public function providerTRIM()
+    {
+        return require 'data/Calculation/TextData/TRIM.php';
+    }
+}
