+++ tests/PhpSpreadsheetTests/Calculation/Functions/TextData/ValueTest.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\TextData;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheet\Calculation\TextData;
+use PhpOffice\PhpSpreadsheet\Shared\StringHelper;
+use PHPUnit\Framework\TestCase;
+
+class ValueTest extends TestCase
+{
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    public function tearDown()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(',');
+        StringHelper::setCurrencyCode('$');
+    }
+
+    /**
+     * @dataProvider providerVALUE
+     *
+     * @param mixed $expectedResult
+     * @param $value
+     */
+    public function testVALUE($expectedResult, $value)
+    {
+        StringHelper::setDecimalSeparator('.');
+        StringHelper::setThousandsSeparator(' ');
+        StringHelper::setCurrencyCode('$');
+
+        $result = TextData::VALUE($value);
+        $this->assertEquals($expectedResult, $result, '', 1E-8);
+    }
+
+    public function providerVALUE()
+    {
+        return require 'data/Calculation/TextData/VALUE.php';
+    }
+}
